+++ .github/workflows/ci.yml
@@ -7,7 +7,7 @@ jobs:
       - uses: actions/checkout@v2.4.0
       - uses: actions/setup-node@v2.4.1
         with:
-          node-version: '14.x'
+          node-version: 12 # https://github.com/actions/runner/issues/772
       - run: npm ci
       - run: npm run lint
       - run: npm run test
